Refactor imagevector package to conform to usage pattern in gardener/gardener
